refactor: testutils into its own crate
