+++ Cargo.toml
@@ -8,6 +8,8 @@ members = [
     "sk-driver",
     "sk-store",
     "sk-tracer",
+
+    "testutils",
 ]
 
 [workspace.package]
@@ -38,7 +40,7 @@ derive_setters = "0.1.6"
 dirs = "5.0.1"
 either = "1.13.0"
 futures = "0.3.31"
-json-patch-ext = "0.2.0"
+json-patch-ext = "0.2.1"
 k8s-openapi = { version = "0.23.0", features = ["v1_27"] }
 lazy_static = "1.5.0"
 object_store = { version = "0.11.1", features = ["aws", "gcp", "azure", "http"] }
@@ -60,11 +62,12 @@ tracing-subscriber = { version = "0.3.18", features = ["env-filter"] }
 url = "2.5.3"
 
 # test dependencies
+sk-testutils = { path = "testutils" }
 http = "1.1.0"
 httpmock = "0.6.8"
 hyper = "1.5.0"
 insta = "1.41.1"
-mockall = "0.11.4"
+mockall = "0.13.1"
 rstest = "0.18.2"
 tracing-test = "0.2.5"
