+++ sk-cli/src/xray/tests/mod.rs
@@ -1,6 +1,6 @@
 mod app_test;
 
 use rstest::*;
+use sk_testutils::*;
 
 use super::app::*;
-use super::*;
