+++ sk-cli/src/xray/view/tests/helpers.rs
