+++ sk-cli/src/xray/view/tests/mod.rs
@@ -1,4 +1,4 @@
-mod testutils;
+mod helpers;
 mod view_test;
 
 use rstest::*;
