+++ sk-ctrl/Cargo.toml
@@ -8,9 +8,6 @@ license-file.workspace = true
 readme.workspace = true
 edition.workspace = true
 
-[features]
-testutils = []
-
 [dependencies]
 anyhow = { workspace = true }
 clap = { workspace = true }
@@ -33,5 +30,5 @@ tracing = { workspace = true }
 [dev-dependencies]
 httpmock = { workspace = true }
 rstest = { workspace = true }
-sk-core = { workspace = true, features = ["testutils"] }
+sk-testutils = { workspace = true }
 tracing-test = { workspace = true }
