+++ sk-ctrl/src/tests/mod.rs
@@ -1,7 +1,8 @@
 mod controller_test;
 
 use rstest::*;
-use sk_core::k8s::testutils::*;
+use sk_core::prelude::*;
+use sk_testutils::*;
 use tracing_test::traced_test;
 
 use super::*;
