+++ sk-driver/Cargo.toml
@@ -8,9 +8,6 @@ license-file.workspace = true
 readme.workspace = true
 edition.workspace = true
 
-[features]
-testutils = []
-
 [dependencies]
 anyhow = { workspace = true }
 clap = { workspace = true }
@@ -34,6 +31,6 @@ httpmock = { workspace = true }
 mockall = { workspace = true }
 rstest = { workspace = true }
 rmp-serde = { workspace = true }
-sk-core = { workspace = true, features = ["testutils"] }
-sk-store = { workspace = true, features = ["testutils"] }
+sk-store = { workspace = true, features = ["mock"] }
+sk-testutils = { workspace = true }
 tracing-test = { workspace = true }
