+++ sk-driver/src/tests/mod.rs
@@ -4,6 +4,7 @@ mod runner_test;
 
 use rstest::*;
 use sk_core::prelude::*;
+use sk_testutils::*;
 use tracing_test::traced_test;
 
 use super::mutation::*;
