+++ sk-store/Cargo.toml
@@ -9,7 +9,7 @@ readme.workspace = true
 edition.workspace = true
 
 [features]
-testutils = ["dep:mockall"]
+mock = ["dep:mockall"]
 
 [dependencies]
 anyhow = { workspace = true }
@@ -27,13 +27,13 @@ sk-core = { workspace = true }
 thiserror = { workspace = true }
 tracing = { workspace = true }
 
-# testutils dependencies
+# mock dependencies
 mockall = { workspace = true, optional = true }
 
 [dev-dependencies]
 assertables = { workspace = true }
 clockabilly = { workspace = true, features = ["mock"] }
 rstest = { workspace = true }
-sk-core = { workspace = true, features = ["testutils"] }
+sk-testutils = { workspace = true }
 tokio = { workspace = true }
 tracing-test = { workspace = true }
