+++ sk-store/src/tests/mod.rs
@@ -3,7 +3,8 @@ mod pod_owners_map_test;
 mod trace_store_test;
 
 use rstest::*;
-use sk_core::k8s::testutils::*;
+use sk_core::prelude::*;
+use sk_testutils::*;
 use tracing_test::traced_test;
 
 use super::*;
