+++ sk-store/src/watchers/tests/mod.rs
@@ -4,6 +4,7 @@ use futures::stream;
 use mockall::predicate;
 use rstest::*;
 use sk_core::prelude::*;
+use sk_testutils::*;
 use tracing_test::traced_test;
 
 use super::*;
