+++ testutils/data/validation_trace.json
@@ -1,0 +1,116 @@
+{
+    "version": 2,
+    "config": {
+        "trackedObjects": {
+            "apps/v1.Deployment": {
+                "podSpecTemplatePath": "/spec/template"
+            }
+        }
+    },
+    "events": [
+        {
+            "ts": 1709241485,
+            "applied_objs": [
+                {
+                    "apiVersion": "apps/v1",
+                    "kind": "Deployment",
+                    "metadata": {
+                        "annotations": {},
+                        "name": "nginx-deployment",
+                        "namespace": "default"
+                    },
+                    "spec": {
+                        "replicas": 2,
+                        "selector": {
+                            "matchLabels": {
+                                "app": "nginx"
+                            }
+                        },
+                        "template": {
+                            "metadata": {
+                                "creationTimestamp": null,
+                                "labels": {
+                                    "app": "nginx"
+                                }
+                            },
+                            "spec": {
+                                "containers": [
+                                    {
+                                        "name": "nginx",
+                                        "image": "nginx:1.14.2",
+                                        "ports": [
+                                            {
+                                                "containerPort": 80,
+                                                "protocol": "TCP"
+                                            }
+                                        ],
+                                        "resources": {},
+                                        "terminationMessagePath": "/dev/termination-log",
+                                        "terminationMessagePolicy": "File",
+                                        "imagePullPolicy": "IfNotPresent"
+                                    }
+                                ],
+                                "restartPolicy": "Always",
+                                "terminationGracePeriodSeconds": 30,
+                                "dnsPolicy": "ClusterFirst",
+                                "securityContext": {},
+                                "schedulerName": "default-scheduler"
+                            }
+                        },
+                        "strategy": {
+                            "type": "RollingUpdate",
+                            "rollingUpdate": {
+                                "maxUnavailable": "25%",
+                                "maxSurge": "25%"
+                            }
+                        },
+                        "revisionHistoryLimit": 10,
+                        "progressDeadlineSeconds": 600
+                    },
+                    "status": {
+                        "observedGeneration": 1,
+                        "replicas": 2,
+                        "updatedReplicas": 2,
+                        "readyReplicas": 2,
+                        "availableReplicas": 2,
+                        "conditions": [
+                            {
+                                "type": "Available",
+                                "status": "True",
+                                "lastUpdateTime": "2024-02-16T17:02:25Z",
+                                "lastTransitionTime": "2024-02-16T17:02:25Z",
+                                "reason": "MinimumReplicasAvailable",
+                                "message": "Deployment has minimum availability."
+                            },
+                            {
+                                "type": "Progressing",
+                                "status": "True",
+                                "lastUpdateTime": "2024-02-16T17:02:25Z",
+                                "lastTransitionTime": "2024-02-16T17:02:23Z",
+                                "reason": "NewReplicaSetAvailable",
+                                "message": "ReplicaSet \"nginx-deployment-86dcfdf4c6\" has successfully progressed."
+                            }
+                        ]
+                    }
+                }
+            ],
+            "deleted_objs": [
+                {
+                    "apiVersion": "apps/v1",
+                    "kind": "Deployment",
+                    "metadata": {
+                        "name": "nginx-deployment-2",
+                        "namespace": "default"
+                    },
+                    "spec": {}
+                }
+            ]
+        }
+    ],
+    "index": {
+        "apps/v1.Deployment": {
+            "default/nginx-deployment": 2842228259284014139
+        }
+    },
+    "pod_lifecycles": {}
+}
