+++ testutils/src/lib.rs
@@ -1,9 +1,13 @@
+mod constants;
 mod fake;
 mod objs;
 mod pods;
 mod sim;
+mod traces;
 
+pub use constants::*;
 pub use fake::*;
 pub use objs::*;
 pub use pods::*;
 pub use sim::*;
+pub use traces::*;
