+++ testutils/src/pods.rs
@@ -1,8 +1,9 @@
 use clockabilly::DateTime;
 use rstest::*;
+use sk_core::macros::*;
+use sk_core::prelude::*;
 
-use crate::macros::*;
-use crate::prelude::*;
+use crate::constants::*;
 
 const CONTAINER_PREFIX: &str = "container";
 const INIT_CONTAINER_PREFIX: &str = "init-container";
